#489 move host/ip resolving code to WebDriverException
to avoid circular dependency between modules "webdriver-api" and "net"
