+++ java/client/src/org/openqa/selenium/remote/RemoteWebDriver.java
@@ -58,7 +58,6 @@ import org.openqa.selenium.logging.LoggingHandler;
 import org.openqa.selenium.logging.LoggingPreferences;
 import org.openqa.selenium.logging.Logs;
 import org.openqa.selenium.logging.NeedsLocalLogs;
-import org.openqa.selenium.net.NetworkUtils;
 import org.openqa.selenium.remote.internal.JsonToWebElementConverter;
 import org.openqa.selenium.remote.internal.WebElementToJsonConverter;
 import org.openqa.selenium.security.Credentials;
@@ -81,7 +80,7 @@ public class RemoteWebDriver implements WebDriver, JavascriptExecutor,
     HasInputDevices, HasCapabilities, TakesScreenshot {
 
   static {
-    NetworkUtils.scheduleIpHostResolving();
+    WebDriverException.scheduleIpHostResolving();
   }
 
   // TODO(dawagner): This static logger should be unified with the per-instance localLogs
